requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
pyodbc
azure-identity
python-dotenv
pydantic